test(qa): split scenario-selection regressions
